[Fleet] Do not allow user to add agent to managed policy
